Fix the exception Error: Chunk.entry was removed. Use hasRuntime() with newer versions of yarn.
The issue was caused by the extract-text-webpack-plugin module v1.0.1,
which uses webpack but lists it only in the devDependencies. The global
installed version of webpack was the 3.X.X (as storybook requires) which
is not compatible with the version 1.0.1 of the extract-text-webpack-plugin.

Using old versions of yarn (<= 1.3.2) everything work as they install
webpack as a submodule of extract-text-webpack-plugin despite being listed
in the devDependencies. Newer versions simply don't work.

This commit sets a global installed version of webpack compatible with
extract-text-webpack-plugin v1.0.1 and uses the selective versions
resolution offered by yarn (see also https://goo.gl/pqdMJf) to install
newer versions for the imcompatible packages required by storybook.